Refactor(routes): move specific routes into respected directories following the flat-routes convention
