+++ app/components/DropDownIcon.tsx
@@ -46,13 +46,13 @@ export default function DropDownIcon({ imageId, username }: { imageId?: string; 
 							<Menu.Item>
 								{({ active }) => (
 									<a
-										href={`/${username}/account`}
+										href={`/${username}/settings`}
 										className={classNames(
 											active ? 'bg-gray-100 text-gray-900' : 'text-gray-700',
 											'block px-4 py-2 text-sm',
 										)}
 									>
-										My Account
+										Account Settings
 									</a>
 								)}
 							</Menu.Item>
