+++ app/routes/$username_+/_index.tsx
@@ -1,0 +1,8 @@
+export default function ProfileRoute() {
+	return (
+		<div className="text-text-primary">
+			<h1>Profile</h1>
+			<p>This is the profile page</p>
+		</div>
+	);
+}
