+++ app/routes/_auth+/logout.tsx
