+++ app/routes/_public+/bar-of-the-month.tsx
