+++ app/routes/_public+/discovery.tsx
